requirements.txt
requests
python-dotenv
pydantic
pydantic-settings
google-generativeai
psycopg2-binary
langchain==0.1.20
langchain-groq==0.1.9
langchain-community==0.0.38
langchain-core==0.1.52
fastapi
uvicorn
